reports + join req protect backend

+++ backend/controllers/posts.js
@@ -140,13 +140,19 @@ const postController = {
         res.sendStatus(200);
     },
     delete: async (req, res) => {
+        const [boarditsClient, Boardits] = await getModelCon("boardits");
+        let mods = await Boardits.find({name: req.body.boarditName}, "mods");
+        mods = mods[0].mods;
+        if (!mods.includes(jwt.decode(req.token).username)) {
+            return res.sendStatus(403);
+        }
+
         const [client, Posts] = await getModelCon("posts");
         await Posts.deleteOne({id: req.params.postID});
 
         const [reportsClient, Reports] = await getModelCon("reports");
         await Reports.deleteOne({reportedPost: req.params.postID});
 
-        const [boarditsClient, Boardits] = await getModelCon("boardits");
         await Boardits.updateOne(
             {name: req.body.boarditName},
             {$pull: {posts: req.params.postID}, $inc: {deletedPosts: 1}}
